src/requirements.txt
aiohttp==2.1.0
kafka-python==1.3.3
pytz==2017.2
requests==2.13.0
retrying==1.3.3
ujson==1.35
uvloop==0.8.0
Scrapy==1.3.3
